Minor fixes to error msgs to CLI if pool/cluster doesn't exist
