+++ clusterman/batch/cluster_metrics_collector.py
@@ -46,13 +46,13 @@ from clusterman.batch.util import suppress_request_limit_exceeded
 from clusterman.config import get_pool_config_path
 from clusterman.config import load_cluster_pool_config
 from clusterman.config import setup_config
+from clusterman.exceptions import ClusterNotFoundError
 from clusterman.mesos.metrics_generators import ClusterMetric
 from clusterman.mesos.metrics_generators import generate_framework_metadata
 from clusterman.mesos.metrics_generators import generate_kubernetes_metrics
 from clusterman.mesos.metrics_generators import generate_simple_metadata
 from clusterman.mesos.metrics_generators import generate_system_metrics
 from clusterman.util import All
-from clusterman.util import ClusterNotFoundError
 from clusterman.util import get_pool_name_list
 from clusterman.util import sensu_checkin
 from clusterman.util import setup_logging
@@ -104,7 +104,7 @@ class ClusterMetricsCollector(BatchDaemon, BatchLoggingMixin, BatchRunningSentin
                 self.pools[scheduler] = get_pool_name_list(self.options.cluster, scheduler)
             except ClusterNotFoundError as e:
                 logger.error(e)
-                raise SystemExit
+                raise
         for scheduler, pools in self.pools.items():
             for pool in pools:
                 watcher_config = {
